feat: overhaul the installaion script
Now the script wont create symlinks for folders, it will create the
folder on the user directory if it doesn't exist and create a symlink
for the files there.

This was a problem previously because the .config folder may already
exist, so the install script could overwrite the folder or just fail to
execute.